Fix me thinking I was clever
When I first started in golang (coming from python), I absolutely despised
golang's error handling. In the intervening years I've come to respect it for
what it is (even if I still think it's a little boilerplatey). This little bit
of cleverness I thought up on day 2 of golang hasn't served me well (it caused
a bug I've been chasing off/on for 2 years). Time to get rid of it.